requirements.txt
flask==3.0.0
tweepy==4.14.0
python-dotenv==1.0.0
pandas==2.1.1
scikit-learn==1.3.2
numpy==1.26.2
requests==2.31.0
joblib==1.3.2
werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3 